requirements.txt
flask
flask-jwt-extended
pymongo
python-dotenv
